+++ src/utils/utils.ts
@@ -174,3 +174,11 @@ export function fixPronunciations(text: string): string {
 
   return text;
 }
+
+/**
+ * Required for the !reload command to work
+ */
+export function requireUncached(module) {
+  delete require.cache[require.resolve(module)];
+  return require(module);
+}